refactor: split Exx_Info struct into independent header files
This is the first step of removing GlobalC::exx_info global variable.

## Changes
- Split Exx_Info into 4 independent header files:
  - exx_info_global.h: Exx_Info_Global (core global settings)
  - exx_info_lip.h: Exx_Info_Lip (LIP method parameters)
  - exx_info_ri.h: Exx_Info_RI (RI method parameters)
  - exx_info_opt_abfs.h: Exx_Info_Opt_ABFs (optimized ABF parameters)
- Modified exx_info.h to include the 4 new headers
- Updated all references from Exx_Info::Exx_Info_* to Exx_Info_*

## Benefits
1. Reduced coupling: modules can now include only needed headers
2. Types are now independent, preparing for parameter passing
3. GlobalC::exx_info remains unchanged, fully backward compatible

## Next Steps (Phase 2-3)
Phase 2: Replace references with value copies in Exx_Info_Lip/RI
  - Exx_Info_Lip: ccp_type& and hse_omega& -> value copies
  - Exx_Info_RI: coulomb_param& -> value copy
  - This removes dependency on Exx_Info_Global lifetime

Phase 3: Remove GlobalC::exx_info global variable
  - Pass Exx_Info_* as parameters instead of global access
  - Update ~60 files that use GlobalC::exx_info
  - Modules: module_xc, module_ri, source_lcao, source_esolver, source_pw
