+++ source/source_lcao/module_ri/Exx_LRI.h
@@ -60,7 +60,7 @@ private:
 	using TatomR = std::array<double,Ndim>;		// tmp
 
 public:
-	Exx_LRI(const Exx_Info::Exx_Info_RI& info_in) :info(info_in) {}
+	Exx_LRI(const Exx_Info_RI& info_in) :info(info_in) {}
 	Exx_LRI operator=(const Exx_LRI&) = delete;
 	Exx_LRI operator=(Exx_LRI&&);
 
@@ -116,7 +116,7 @@ public:
 private:
 	// WARNING: reference to Exx_Info_RI, which holds references into Exx_Info_Global.
 	// Must not outlive GlobalC::exx_info. See exx_info.h for details.
-	const Exx_Info::Exx_Info_RI &info;
+	const Exx_Info_RI &info;
 	MPI_Comm mpi_comm;
 	const K_Vectors *p_kv = nullptr;
 	std::shared_ptr<ORB_gaunt_table> MGT;
