+++ source/source_lcao/module_ri/exx_abfs-jle.h
@@ -2,7 +2,7 @@
 #define EXX_ABFS_JLE_H
 
 #include "exx_abfs.h"
-#include "../../source_hamilt/module_xc/exx_info.h"
+#include "source_hamilt/module_xc/exx_info_opt_abfs.h"
 #include "../../source_basis/module_ao/ORB_atomic_lm.h"
 
 #include <vector>
@@ -15,7 +15,7 @@ class Exx_Abfs::Jle
 public:
 	static std::vector<std::vector<std::vector<Numerical_Orbital_Lm>>>
 	init_jle(
-		const Exx_Info::Exx_Info_Opt_ABFs &info,
+		const Exx_Info_Opt_ABFs &info,
 		const double kmesh_times, 
 		const UnitCell& ucell,
 		const LCAO_Orbitals& orb);
