+++ source/source_lcao/module_ri/exx_lip.hpp
@@ -84,7 +84,7 @@ void Exx_Lip<T, Device>::cal_exx()
 }
 
 template <typename T, typename Device>
-Exx_Lip<T, Device>::Exx_Lip(const Exx_Info::Exx_Info_Lip& info_in,
+Exx_Lip<T, Device>::Exx_Lip(const Exx_Info_Lip& info_in,
                             const ModuleSymmetry::Symmetry& symm,
                             K_Vectors* kv_ptr_in,
                             psi::Psi<T, Device>* psi_local_in,
