+++ source/source_lcao/module_ri/exx_rotate_abfs.h
@@ -3,6 +3,7 @@
 #define EXX_ROTATE_ABFS_H
 
 #include "LRI_CV.h"
+#include "source_hamilt/module_xc/exx_info_ri.h"
 // #include "module_xc/exx_info.h"
 // #include "module_basis/module_ao/ORB_atomic_lm.h"
 #include "Exx_LRI.h"
@@ -27,7 +28,7 @@ class Moment_abfs
     using TAq = std::pair<TA, Tq>;
 
   public:
-    Moment_abfs(Exx_Info::Exx_Info_RI& info_in) : info(info_in) {};
+    Moment_abfs(Exx_Info_RI& info_in) : info(info_in) {};
     ~Moment_abfs() {};
     void cal_VR(
         const UnitCell& ucell,
@@ -69,7 +70,7 @@ class Moment_abfs
     // std::map<TA, std::map<TAC, RI::Tensor<Tdata>>> VR;
     // WARNING: reference to Exx_Info_RI, which holds references into Exx_Info_Global.
     // Must not outlive GlobalC::exx_info. See exx_info.h for details.
-    Exx_Info::Exx_Info_RI& info;
+    Exx_Info_RI& info;
 };
 #include "exx_rotate_abfs.hpp"
